adding replication manager back to server app due to high coupling
