+++ Instant-Messenger/include/server/ReplicationManager.hpp
@@ -16,21 +16,7 @@ using namespace servergroupmanager;
 
 class ReplicationManager : public Socket{
     public:
-        ReplicationManager(ServerGroupManager *groupManager);
-        void connectToRmServers();
-        std::map<int, sockaddr_in> rm_connect_sockets_fd;
-        void createRMListenerSocket();
-        static void *handleRMCommunication(void *args);
-        static void *acceptRMConnection(void *args);
-        int rmNumber;
-        int rm_listening_socket_fd;
-        struct sockaddr_in rm_listening_serv_addr;
-        void printRMConnections() const;
-        Semaphore* sockets_connections_semaphore;
-        void setRmNumber(int rmNumber);
-        int getRmNumber();
-        void sendMockDataToRMServers();
-        ServerGroupManager *groupManager;
+
 
     private:
