+++ Instant-Messenger/include/util/Packet.hpp
@@ -87,8 +87,8 @@ struct Packet
     bool isJoinMessage() {
         return this->type == JOIN_PACKET;
     }
-    bool isReplicationMessage() {
-        return this->type == REPLICATION_MESSAGE;
+    bool isReplicationPacket() {
+        return this->type == REPLICATION_PACKET;
     }
 };
