+++ Instant-Messenger/include/util/definitions.hpp
@@ -63,7 +63,8 @@ enum
     CONNECTION_REFUSED_PACKET,
     DISCONNECT_PACKET,
     ELECTION_PACKET,
-    REPLICATION_MESSAGE
+    REPLICATION_PACKET,
+    REPLICATION_CONFIRMATION_PACKET
 };
 ////////
 #endif
