+++ Instant-Messenger/src/server/server.cpp
@@ -1,7 +1,6 @@
 #include "../../include/server/server.hpp"
 #include "../../include/util/definitions.hpp"
 #include "../../include/client/ConnectionKeeper.hpp"
-#include "../../include/server/ReplicationManager.hpp"
 
 #ifdef _WIN32
 #include <Windows.h>
@@ -20,7 +19,12 @@ namespace server {
         sockets_connections_semaphore = new Semaphore(1);
         groupManager = new ServerGroupManager();
         connectionMonitor = new ConnectionMonitor();
-        replicationManager = new ReplicationManager(groupManager);
+
+        rm_listening_serv_addr.sin_family = AF_INET;
+        rm_listening_serv_addr.sin_addr.s_addr = INADDR_ANY;
+        bzero(&(rm_listening_serv_addr.sin_zero), 8);
+
+        rm_listening_socket_fd = 0;
 
         // TODO assim como proxyFE, o server vai ter que ter 2 sockets, um pra conectar nos FE e outro nos RM
         {
@@ -36,14 +40,14 @@ namespace server {
 
 
     void Server::prepareReplicationManager(int rmNumber) {
-        replicationManager->setRmNumber(rmNumber);
-        replicationManager->connectToRmServers();
+        this->setRmNumber(rmNumber);
+        this->connectToRmServers();
 
-        if(replicationManager->getRmNumber() > 0) {
-            replicationManager->createRMListenerSocket();
+        if(this->getRmNumber() > 0) {
+            this->createRMListenerSocket();
         }
         else {
-            replicationManager->sendMockDataToRMServers();
+            this->sendMockDataToRMServers();
         }
     }
 
@@ -146,7 +150,12 @@ namespace server {
             Packet *connectionRefusedPacket = new Packet(CONNECTION_REFUSED_PACKET);
             strcpy(connectionRefusedPacket->user_id, registrationPacket->user_id);
             std::cout << "[DEBUG|ERROR] limit sessoes alcanadas pelo user" << std::endl;
-            this->sendPacket(frontEndSocket, connectionRefusedPacket);
+
+            //todo: do we need to know if on RM server the connection was refused?
+            if(!registrationPacket->isReplicationPacket()) {
+                this->sendPacket(frontEndSocket, connectionRefusedPacket);
+            }
+
             this->sockets_connections_semaphore->post();
             return -1;
         }
@@ -217,6 +226,8 @@ namespace server {
     void *Server::listenFrontEndCommunication(void *args) {
         Server* _this = (Server *) args;
         bool connectedFrontEnd = true;
+        bool replicationError = false;
+
         while (connectedFrontEnd) {
             // Listen for an incoming Packet from client
             Packet *receivedPacket = _this->readPacket(_this->socket_fd, &connectedFrontEnd);
@@ -224,6 +235,28 @@ namespace server {
                 // Free allocated memory for reading Packet
                 free(receivedPacket);
                 break;
+            }
+
+            // verify if should send replication socket or not
+            if(_this->getIsPrimaryServer() && (receivedPacket->isMessage() || receivedPacket->isJoinMessage() || receivedPacket->isDisconnect())) {
+                cout << "Sending " << _this->rm_connect_sockets_fd.size() << " replication packets" << endl;
+
+                for(auto socket : _this->rm_connect_sockets_fd) {
+                    cout << "Sending packet of type " << receivedPacket->type << " to socket " << socket.first << endl;
+                    _this->sendPacket(socket.first, receivedPacket); // send message for each socket on RM socket list
+                    Packet *confirmationPacket = _this->readPacket(socket.first, &connectedFrontEnd); // wait for socket answer
+                    cout << "Reading packet of type " << confirmationPacket->type << " on socket " << socket.first << endl;
+
+                    if (!connectedFrontEnd) {
+                        // Free allocated memory for reading Packet
+                        free(receivedPacket);
+                        replicationError = true;
+                    }
+                }
+            }
+
+            if(replicationError) {
+                cout << "Error on replicating message. What should we do?" << endl;
             }
 
             if (receivedPacket->isMessage()) {
@@ -244,17 +277,6 @@ namespace server {
                 connectionId.first = receivedPacket->clientDispositiveIdentifier;
                 connectionId.second = _this->socket_fd;
                 _this->closeClientConnection(connectionId); // considers the front end connection
-            } /*else if (receivedPacket->isReplicationMessage()) {
-                cout << "[DEBUG] recebi replicacao" << receivedPacket->message << endl;
-                _this->groupManager->processReceivedPacket(receivedPacket);
-            }*/
-
-            // verify if should send replication socket or not
-            if(_this->getIsPrimaryServer() && (receivedPacket->isMessage() || receivedPacket->isJoinMessage() || receivedPacket->isDisconnect())) {
-                for(auto socket : _this->rm_connect_sockets_fd) {
-                    receivedPacket->type = REPLICATION_MESSAGE;
-                    _this->sendPacket(socket.first, receivedPacket); // send message for each socket on RM socket list
-                }
             }
         }
 
@@ -312,5 +334,218 @@ namespace server {
     void Server::setIsPrimaryServer(bool value)
     {
         this->isPrimaryServer = value;
+    }
+
+
+    void Server::connectToRmServers() {
+        for(int connectMachineNumber = rmNumber; connectMachineNumber < MAX_RM; connectMachineNumber++) {
+            int *connectSocket = (int *) calloc(1, sizeof(int *));
+            char ip_address[10] = "127.0.0.1";
+            struct sockaddr_in rm_connect_socket_addr{};
+
+            rm_connect_socket_addr.sin_family = AF_INET;
+            rm_connect_socket_addr.sin_port = htons(RM_BASE_PORT_NUMBER + connectMachineNumber + 1); // port is base port + next machine RM number (rmNumber+1)
+            bzero(&(rm_connect_socket_addr.sin_zero), 8);
+
+            if (inet_pton(AF_INET, ip_address, &rm_connect_socket_addr.sin_addr) <= 0) {
+                cout << "\n RM Invalid address/ Address not supported \n" << endl;
+            }
+
+            if ((*connectSocket = socket(AF_INET, SOCK_STREAM, 0)) < 0) {
+                cout << "\n RM Socket creation error \n" << endl;
+            }
+
+            if (connect(*connectSocket, (struct sockaddr *) &rm_connect_socket_addr, sizeof(rm_connect_socket_addr)) < 0) {
+                cout << "ERROR connecting on RM sockets\n" << endl;
+            }
+
+            rm_connect_sockets_fd.insert({*connectSocket, rm_connect_socket_addr});
+            cout << "Connected to socket " << *connectSocket << " and port " << ntohs(rm_connect_socket_addr.sin_port) << endl;
+
+            std::pair<int *, Server *> *args = (std::pair<int *, Server *> *) calloc(1, sizeof(std::pair<int *, Server *>));
+
+            // Send pointer of the previously allocated address and be able to access it's value in new thread's execution
+            args->first = connectSocket;
+            args->second = this;
+
+            pthread_t connectedRMThread;
+            pthread_create(&connectedRMThread, NULL, handleRMCommunication, (void *) args);
+        }
+    }
+
+/**
+* This method creates listener sockets based on RM number
+*/
+    void Server::createRMListenerSocket() {
+        int opt = 1;
+        int rmListeningPort = RM_BASE_PORT_NUMBER + rmNumber;
+        rm_listening_serv_addr.sin_port = htons(rmListeningPort);
+
+        //cout << "Listening on port " << rmListeningPort << endl;
+
+        // Create socket file descriptor
+        if ((rm_listening_socket_fd = socket(AF_INET, SOCK_STREAM, 0)) <= 0) {
+            cout << "ERROR opening socket\n" << endl;
+            exit(1);
+        }
+
+        // Forcefully attaching socket to the port
+        if (setsockopt(rm_listening_socket_fd, SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(opt)))
+        {
+            perror("setsockopt");
+            exit(EXIT_FAILURE);
+        }
+
+        // Attach socket to server's port
+        if (::bind(rm_listening_socket_fd, (struct sockaddr *) &rm_listening_serv_addr, sizeof(rm_listening_serv_addr)) < 0) {
+            cout << "ERROR on binding\n" << endl;
+            exit(1);
+        }
+
+        // Configure socket to listen for tcp connections
+        if (::listen(rm_listening_socket_fd, MAXBACKLOG) < 0) // SOMAXCONN is the maximum value of backlog
+        {
+            cout << "ERROR on listening\n" << endl;
+            exit(1);
+        }
+
+        cout << "\nSetting listener socket for RM machine of number " << rmNumber << endl;
+        cout << "Listening for RM connections on socket: " << rm_listening_socket_fd << " and port " << ntohs(
+                rm_listening_serv_addr.sin_port) << endl;
+
+        pthread_t acceptRMConnectionThread;
+        pthread_create(&acceptRMConnectionThread, NULL, acceptRMConnection, (void *) this);
+    }
+
+    void *Server::acceptRMConnection(void *param) {
+        while(1) {
+            int *newsockfd = (int *) calloc(1, sizeof(int *));
+            socklen_t clilen = sizeof(struct sockaddr_in);
+
+            // reference to this class
+            Server *_this;
+            _this = (Server *) param;
+
+            if ((*newsockfd = accept(_this->rm_listening_socket_fd, (struct sockaddr *) &_this->rm_listening_serv_addr,
+                                     &clilen)) == -1) {
+                cout << "ERROR on accept\n" << endl;
+                cout << "Could not connect to port " << _this->rm_listening_serv_addr.sin_port << endl;
+            }
+            else {
+                std::cout << "Máquina conectada pelo socket " << *newsockfd << " de porta " << ntohs(_this->rm_listening_serv_addr.sin_port) << std::endl;
+
+                std::pair<int *, Server *> *args = (std::pair<int *, Server *> *) calloc(1, sizeof(std::pair<int *, Server *>));
+                _this->rm_connect_sockets_fd.insert({*newsockfd, _this->rm_listening_serv_addr});
+
+                cout << "### Vetor de sockets de Replicacao ###" << endl;
+                _this->printRMConnections();
+                cout << "######################################" << endl;
+
+                // Send pointer of the previously allocated address and be able to access it's value in new thread's execution
+                args->first = newsockfd;
+                // Also, send reference of this instance to the new thread
+                args->second = _this;
+
+                pthread_t listenRMThread;
+                pthread_create(&listenRMThread, NULL, handleRMCommunication, (void *) args);
+            }
+        }
+    }
+
+    void *Server::handleRMCommunication(void *args)
+    {
+        // We cast our receve id void* args to a pair*
+        std::pair<int *, Server *> *args_pair = (std::pair<int *, Server *> *) args;
+
+        // Read socket pointer's value and free the previously allocated memory in the main thread
+        int rm_socket_fd = *(int *) args_pair->first;
+        free(args_pair->first);
+
+        // Create a reference of the instance in this thread
+        Server *_this = (Server*)calloc(1, sizeof(Server*));
+        _this = (Server *) args_pair->second;
+
+        // Free pair created for sending arguments
+        free(args_pair);
+
+        bool connectedClient = true;
+        while (connectedClient)
+        {
+            _this->sockets_connections_semaphore->wait();
+            cout << "[Communication Thread] - Waiting socket " << rm_socket_fd << " messages" << endl;
+            _this->sockets_connections_semaphore->post();
+            // Listen for an incoming Packet from client
+            // todo: we will receive messages here and we need to process them accordingly
+            Packet *receivedPacket = _this->readPacket(rm_socket_fd, &connectedClient);
+            cout << "Received packet from socket " << rm_socket_fd << " of type " << receivedPacket->type << endl;
+            if (!connectedClient)
+            {
+                // Free allocated memory for reading Packet
+                free(receivedPacket);
+                break;
+            }
+
+            //create field on socket to store customer socket id
+            /*if (receivedPacket->isMessage()) {
+                receivedPacket->type = REPLICATION_PACKET;
+                _this->groupManager->processReceivedPacket(receivedPacket);
+            } else if (receivedPacket->isJoinMessage()) {
+                receivedPacket->type = REPLICATION_PACKET;
+                _this->registerUserToServer(receivedPacket, receivedPacket->frontEndSocket); // considers the front end connection
+            } else if (receivedPacket->isDisconnect()) {
+                receivedPacket->type = REPLICATION_PACKET;
+                pair<int, int> connectionId = pair<int, int>();
+                connectionId.first = receivedPacket->clientDispositiveIdentifier;
+                connectionId.second = receivedPacket->clientSocket;
+                _this->closeClientConnection(connectionId); // considers the front end connection
+            }*/
+
+            cout << "Sending replication packet confirmation to Primary Server from socket " << rm_socket_fd << endl;
+            Packet *confirmationPackage = new Packet();
+            confirmationPackage->type = REPLICATION_CONFIRMATION_PACKET;
+            _this->sendPacket(rm_socket_fd, confirmationPackage);
+        }
+
+        auto socket_it = _this->rm_connect_sockets_fd.find(rm_socket_fd);
+        _this->rm_connect_sockets_fd.erase(socket_it);
+
+        close(rm_socket_fd);
+        return 0;
+    }
+
+    void Server::sendMockDataToRMServers() {
+        sockets_connections_semaphore->wait();
+        bool connectedFrontEnd = true;
+
+        cout << "Sending " << rm_connect_sockets_fd.size() << " replication packets" << endl;
+        for ( const pair<int, struct sockaddr_in> &connectedMachine : rm_connect_sockets_fd)
+        {
+            Packet *mockPacket = new Packet();
+            mockPacket->type = REPLICATION_PACKET;
+            cout << "Sending packet of type " << mockPacket->type << " to socket " << connectedMachine.first << " on port " << ntohs(connectedMachine.second.sin_port) << endl ;
+            sendPacket(connectedMachine.first, mockPacket);
+            Packet *confirmationPacket = readPacket(connectedMachine.first, &connectedFrontEnd); // wait for socket answer
+            cout << "Reading packet of type " << confirmationPacket->type << " on socket " << connectedMachine.first << " on port " << ntohs(connectedMachine.second.sin_port) << endl;
+        }
+        cout << "All packets were replicate successfully" << endl;
+        sockets_connections_semaphore->post();
+    }
+
+    void Server::printRMConnections() const {
+        sockets_connections_semaphore->wait();
+        cout << " rm sockets map size " << rm_connect_sockets_fd.size() << endl;
+        for ( const pair<int, struct sockaddr_in> &connectedMachine : rm_connect_sockets_fd)
+        {
+            cout << "Conectado ao servidor RM de porta " << ntohs(connectedMachine.second.sin_port) << " e socket "
+                 << connectedMachine.first << endl;
+        }
+        sockets_connections_semaphore->post();
+    }
+
+    void Server::setRmNumber(int _rmNumber) {
+        this->rmNumber = _rmNumber;
+    }
+    int Server::getRmNumber() {
+        return this->rmNumber;
     }
 }
